# strict/swagger.json
{
    "openapi": "3.0.3",
    "paths": {
        "/event/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"{\\\"where\\\":{},\\\"relations\\\":[]}\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "EventController"
                ]
            }
        },
        "/event/id/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "EventController"
                ]
            }
        },
        "/event/": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "EventController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "createdAt": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "DeletedAt": {
                                        "type": "number"
                                    },
                                    "userId": {
                                        "type": "number"
                                    },
                                    "user": {
                                        "properties": {
                                            "id": {
                                                "type": "number"
                                            },
                                            "username": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "media": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "eventId": {
                                                    "type": "number"
                                                },
                                                "fileInfo": {
                                                    "type": "any"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "EventController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "createdAt": {
                                        "properties": {
                                            "getVarDate": {
                                                "properties": {}
                                            }
                                        }
                                    },
                                    "DeletedAt": {
                                        "type": "number"
                                    },
                                    "userId": {
                                        "type": "number"
                                    },
                                    "user": {
                                        "properties": {
                                            "id": {
                                                "type": "number"
                                            },
                                            "username": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "media": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "eventId": {
                                                    "type": "number"
                                                },
                                                "fileInfo": {
                                                    "type": "any"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/event/{id}": {
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "EventController"
                ]
            }
        },
        "/media/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"{\\\"where\\\":{},\\\"relations\\\":[]}\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MediaController"
                ]
            }
        },
        "/media/id/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MediaController"
                ]
            }
        },
        "/media/": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MediaController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "eventId": {
                                        "type": "number"
                                    },
                                    "fileInfo": {
                                        "type": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MediaController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "eventId": {
                                        "type": "number"
                                    },
                                    "fileInfo": {
                                        "type": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/media/{id}": {
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "MediaController"
                ]
            }
        },
        "/user/{option}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "option",
                        "schema": {
                            "type": "object",
                            "example": "\"{\\\"where\\\":{},\\\"relations\\\":[]}\""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "UserController"
                ]
            }
        },
        "/user/id/{id}": {
            "get": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "UserController"
                ]
            }
        },
        "/user/": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "UserController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "UserController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "UserController"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "delete": {
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "tags": [
                    "UserController"
                ]
            }
        }
    }
}